+++ src/main/java/com/structurizr/dsl/StructurizrDslParser.java
@@ -793,11 +793,18 @@ public final class StructurizrDslParser extends StructurizrDslTokens {
                         new ImageViewContentParser(restricted).parseImage(getContext(ImageViewDslContext.class), dslFile, tokens);
 
                     } else if (inContext(DynamicViewDslContext.class)) {
-                        new DynamicViewContentParser().parseRelationship(getContext(DynamicViewDslContext.class), tokens);
+                        RelationshipView relationshipView = new DynamicViewContentParser().parseRelationship(getContext(DynamicViewDslContext.class), tokens);
 
                         if (inContext(DynamicViewParallelSequenceDslContext.class)) {
                             getContext(DynamicViewParallelSequenceDslContext.class).hasRelationships(true);
                         }
+
+                        if (shouldStartContext(tokens)) {
+                            startContext(new DynamicViewRelationshipContext(relationshipView));
+                        }
+
+                    } else if (URL_TOKEN.equalsIgnoreCase(firstToken) && inContext(DynamicViewRelationshipContext.class)) {
+                        new DynamicViewRelationshipParser().parseUrl(getContext(DynamicViewRelationshipContext.class), tokens.withoutContextStartToken());
 
                     } else if (THEME_TOKEN.equalsIgnoreCase(firstToken) && (inContext(ViewsDslContext.class) || inContext(StylesDslContext.class))) {
                         new ThemeParser().parseTheme(getContext(), tokens);